+++ src/main/java/com/landvibe/landlog/controller/MemberController.java
@@ -1,12 +1,13 @@
 package com.landvibe.landlog.controller;
 
+import com.landvibe.landlog.controller.form.MemberJoinForm;
+import com.landvibe.landlog.controller.form.MemberLoginForm;
 import com.landvibe.landlog.domain.Member;
 import com.landvibe.landlog.service.MemberService;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
