+++ src/main/java/com/landvibe/landlog/repository/MemoryMemberRepository.java
@@ -1,6 +1,6 @@
 package com.landvibe.landlog.repository;
 
-import com.landvibe.landlog.controller.MemberLoginForm;
+import com.landvibe.landlog.controller.form.MemberLoginForm;
 import com.landvibe.landlog.domain.Member;
 
 import org.springframework.stereotype.Repository;
