+++ src/main/java/com/landvibe/landlog/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.landvibe.landlog.service;
 
-import com.landvibe.landlog.controller.MemberLoginForm;
+import com.landvibe.landlog.controller.form.MemberLoginForm;
 import com.landvibe.landlog.domain.Member;
 import com.landvibe.landlog.repository.MemberRepository;
