+++ src/test/java/com/landvibe/landlog/service/MemberServiceTest.java
@@ -1,6 +1,6 @@
 package com.landvibe.landlog.service;
 
-import com.landvibe.landlog.controller.MemberLoginForm;
+import com.landvibe.landlog.controller.form.MemberLoginForm;
 import com.landvibe.landlog.domain.Member;
 import com.landvibe.landlog.repository.MemoryMemberRepository;
 import org.junit.jupiter.api.AfterEach;
